requirements.txt
python-dotenv==0.19.2
requests==2.26.0
openai==0.27.0
beautifulsoup4==4.10.0
Markdown==3.3.6
python-slugify==5.0.2
Pillow==9.0.0
pytest==7.1.2
flake8==4.0.1
black==22.3.0
